+++ src/refactoring/components/admin/AdminProductList.tsx
@@ -2,6 +2,8 @@ import { useState } from "react";
 import type { Product, Discount } from "../../types";
 import { NewProductForm } from "./product/NewProductForm";
 import { UpdateProductForm } from "./product/UpdateProductForm";
+import { ProductToggleButton } from "./product/ProductToggleButton";
+import { ProductList } from "./product/ProductList";
 
 interface AdminProductListProps {
   products: Product[];
@@ -14,28 +16,7 @@ export const AdminProductList = ({
   onProductUpdate,
   onProductAdd,
 }: AdminProductListProps) => {
-  const [editingProduct, setEditingProduct] = useState<Product | null>(null);
   const [showNewProductForm, setShowNewProductForm] = useState(false);
-  const [openProductIds, setOpenProductIds] = useState<Set<string>>(new Set());
-
-  const toggleProductAccordion = (productId: string) => {
-    setOpenProductIds((prev) => {
-      const newSet = new Set(prev);
-      if (newSet.has(productId)) {
-        newSet.delete(productId);
-      } else {
-        newSet.add(productId);
-      }
-      return newSet;
-    });
-  };
-
-  // 새로운 핸들러 함수 추가
-
-  // handleEditProduct 함수 수정
-  const handleEditProduct = (product: Product) => {
-    setEditingProduct({ ...product });
-  };
 
   return (
     <div>
@@ -53,57 +34,7 @@ export const AdminProductList = ({
           setShowNewProductForm={setShowNewProductForm}
         />
       )}
-      <div className="space-y-2">
-        {products.map((product, index) => (
-          <div
-            key={product.id}
-            data-testid={`product-${index + 1}`}
-            className="bg-white p-4 rounded shadow"
-          >
-            <button
-              type="button"
-              data-testid="toggle-button"
-              onClick={() => toggleProductAccordion(product.id)}
-              className="w-full text-left font-semibold"
-            >
-              {product.name} - {product.price}원 (재고: {product.stock})
-            </button>
-            {openProductIds.has(product.id) && (
-              <div className="mt-2">
-                {editingProduct && editingProduct.id === product.id ? (
-                  <UpdateProductForm
-                    products={products}
-                    product={product}
-                    onProductUpdate={onProductUpdate}
-                    editingProduct={editingProduct}
-                    setEditingProduct={setEditingProduct}
-                  />
-                ) : (
-                  <div>
-                    {product.discounts.map((discount, index) => (
-                      // biome-ignore lint/suspicious/noArrayIndexKey: <explanation>
-                      <div key={index} className="mb-2">
-                        <span>
-                          {discount.quantity}개 이상 구매 시{" "}
-                          {discount.rate * 100}% 할인
-                        </span>
-                      </div>
-                    ))}
-                    <button
-                      type="button"
-                      data-testid="modify-button"
-                      onClick={() => handleEditProduct(product)}
-                      className="bg-blue-500 text-white px-2 py-1 rounded hover:bg-blue-600 mt-2"
-                    >
-                      수정
-                    </button>
-                  </div>
-                )}
-              </div>
-            )}
-          </div>
-        ))}
-      </div>
+      <ProductList products={products} onProductUpdate={onProductUpdate} />
     </div>
   );
 };
